Align the SPRT scoring's Cn with the paper and with MAGSACScoring
2fab419 moved MAGSACScoring to the paper's C(n) = (2^(n/2) Gamma(n/2))^-1 but
left MAGSACSPRTScoring computing (1 / 2^(n/2)) * Gamma(n/2). The two forms differ
by Gamma(n/2)^2, so the sibling classes disagreed for three, five and six degrees
of freedom -- ratios 0.785, 1.767 and 4.0 -- while agreeing at two and four where
Gamma(n/2) is 1.

Unreachable today, since every estimator reports two degrees of freedom, which is
why the regression is byte-identical: six scoring types across both use_sprt
settings and seven local optimizers, comparing median geometric error, inlier
counts and summed scores. Worth fixing anyway so the two classes cannot drift
apart the moment another DOF becomes reachable.

+++ include/scoring/magsac_sprt_scoring.h
@@ -239,7 +239,10 @@ class MAGSACSPRTScoring : public AbstractScoring {
     dofIndex_ = degreesOfFreedom - 2;  // Cache DOF index for lookup table optimization
     gammaTable_ = interleavedGammaLookupTable(dofIndex_);  // Interleaved (lower, upper) row
     k = getK(degreesOfFreedom);
-    Cn = 1.0 / std::pow(2.0, degreesOfFreedom / 2.0) * boost::math::tgamma(degreesOfFreedom / 2.0);
+    // 卡方分布的归一化常数，与论文的 C(n) = (2^(n/2) Γ(n/2))^{-1} 一致。
+    // 此前写作 (1 / 2^(n/2)) * Γ(n/2)，与 MAGSACScoring 相差 Γ(n/2)² 倍。
+    Cn = 1.0 / (std::pow(2.0, static_cast<double>(degreesOfFreedom) / 2.0) *
+                boost::math::tgamma(static_cast<double>(degreesOfFreedom) / 2.0));
     squaredSigmaMax = threshold * threshold;
     squaredSigmaMaxPerTwo = squaredSigmaMax / 2.0;
     squaredSigmaMaxPerFour = squaredSigmaMaxPerTwo / 2.0;
